requirements.txt
Pillow==8.2.0
pymongo==3.11.3
regex==2021.4.4
requests==2.25.1
seaborn==0.11.1
tqdm==4.60.0
django
djangorestframework
pyvi
joblib